src/auth/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
PyJWT==2.9.0
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1